refactor: split default mode implementation
